Commit Title: Refactor Signup DID Identifier
This commit refactors the code related to the DID Identifier on the Signup page. It updates the code to use the getId() method instead of the iD property to retrieve the DID Identifier. It also updates the code to use the getId() method when copying the DID Identifier to the clipboard.