Feature/ipfs encryption (#27)
* Update README.md, app/encoding.go, x/identity/protocol/vault/account/account.go, x/identity/protocol/vault/account/internal/mpc/signature.go, and x/identity/types/vault/v1/share.go

This commit updates multiple files to improve the security and functionality of the application. It adds a license badge to the README.md, updates the encoding config, adds signer data to the wallet account, and updates the signature and share configs.

* Configure account to return Tx Signing objects

* Commit Title: Refactor Account and Share Configs

This commit refactors the Account and Share Configs to improve the code structure and make it easier to use. It updates the PubKey() method to use the GetCryptoPubKey() method, and adds a PublicPoint() method to the ShareConfig. Additionally, it adds a CreatedAt field to the AccountConfig and updates the NewAccountConfigFromShares() method to include the Bech32Prefix.

* Closes SNR-86 and SNR-88

* Add OrbitDB to Node

* Add Berty Session Sync

* Introduce protocol context for storing node config values

* Set Webauthn instance directly for session

* Adjust protocol context

* Update auth/session configuration

* Remove Sync from BeginRegistration method

* Remove node from session

* Introduce Wallet Dispatcher for Generation Jobs

* Introduce UCAN Token Creation

* Add WebAuthn PubKey COSE Extraction

* Remove Verification Method Builder Helpers

* Update Webauthn credential storage

* Shift grpc query

* Update Readme

* Closes SNR-132

* Closes SNR-129, SNR-125, and SNR-123

* Refactor Session

* Configure Bond Denom to be uSNR

* Updated genesis

* Refactor types

* Move Node interface to common package

* Operate Wallet Generation over DID Document

* Update configuration files

Updated the Taskfile.yml to include the --home flag, updated the pkg/common/context.go to update the RepoPath, and removed the webauthn functions from x/identity/module.go.

* Update Ignite Version

* Update Dispatcher

* Commit title: Refactor DID Document Verification Methods

This commit refactors the DID Document Verification Methods to make them more consistent and easier to use. It updates the ID field to be lowercase and changes the VerificationMethod, VerificationRelationship, and Service fields to be repeated fields. Additionally, the Services and VerificationMethods messages have been removed.

* Add Workflow Files

* Closes SNR-133

* Closes SNR-133

* Commit title: Update go.mod and go.sum

This commit updates the go.mod and go.sum files to reflect the current dependencies of the project. This includes removing the github.com/cosmos/ibc-go/v6 dependency and adding the lukechampine.com/blake3 dependency. Additionally, some of the existing dependencies have been updated to their latest versions.

* Standardize Types begin implementation of Fine-Tuned Accounts

* Move ante to pkg/crypto

* Fix missed error check

---------

Signed-off-by: Prad Nukala <pradn@sonr.io>