Feature/ucan (#28)
* Connect go

* Update Vault/Auth and Blockchain Dependencies
Update the dependencies for Vault/Auth and Blockchain to the latest versions. This includes updating the buf.gen.yaml file to comment out the bufbuild/connect-go and bufbuild/protocolbuffers/go plugins, and updating the go.mod and go.sum files to reflect the new versions. Additionally, the generate.sh script has been updated to copy the generated files to the correct locations, and the controller.go file has been updated to add options to the New function.

* Cleaned deprecated code

* Fix missed error check

* Integrate filestore

* Add support for blockchain accounts in DIDController

This commit adds support for blockchain accounts in the DIDController. It updates the PubKey.Address() method to use the Address() method of the Address type, adds a CreateAccount() method to the DIDController that returns a VerificationMethod, and adds a GetVerificationMethodByBlockchainAccountID() method to the DidDocument type. Additionally, it updates the Dispatcher tests to use the new methods.

* Closes SNR-145, SNR-100, SNR-146, SNR-128