chore(.goreleaser.yaml): add nfpms configuration for Sonr Binary Node package
The `.goreleaser.yaml` file was modified to add a new configuration for the `nfpms` section. This configuration adds support for creating a Debian package for the Sonr Binary Node. The package information includes the ID, description, homepage, license, and supported formats. This change was made to enable the creation of the Debian package for the Sonr Binary Node.