feat(query.proto): add Query/ServiceRecord RPC method to retrieve service record for a given origin
The `Query/ServiceRecord` RPC method is added to the `Query` service in the `query.proto` file. This method allows retrieving the service record for a given origin. The method is defined with the following details:
- Request type: `QueryServiceRecordRequest` with a required `origin` field that defines the origin of the service.
- Response type: `QueryServiceRecordResponse` with a required `service_record` field that defines the service record for the given origin.
- HTTP mapping: The method is mapped to a GET request at the `/sonrhq/service/v1/service_record/{origin}` endpoint.

This addition enhances the functionality of the `Query` service by providing the ability to retrieve service records based on the origin.