feat: remove controller package
The changes remove the `controller` package from the codebase. The `controller` package was previously used to handle the signing and refreshing of the keyshares, but this functionality has now been moved directly into the `vault` package. The removal of the `controller` package simplifies the codebase and eliminates the need for a separate controller component.